TDI-33997 bug: [5.6.2]Talend Studio - SVN checkout command is not working as expected for Routes https://jira.talendforge.org/browse/TDI-33997
Fix NoClassDefFoundError for commandline
